requirements.txt
asgiref==3.6.0
asn1crypto==1.5.1
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
scramp==1.4.4
six==1.16.0
sqlparse==0.4.3
